requirements.txt
opencv-python
numpy
torch
torchvision
Pillow
easyOCR
opencv-python
streamlit
ultralytics
